renamed original inventory to item and split the orginal code in the respective .h and .cpp files
