+++ Data Structures Game/Data Structures Game.vcxproj
@@ -129,12 +129,12 @@
   <ItemGroup>
     <ClCompile Include="Data Structures Game.cpp" />
     <ClCompile Include="enemy.cpp" />
-    <ClCompile Include="inventory.cpp" />
+    <ClCompile Include="item.cpp" />
     <ClCompile Include="user.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="enemy.h" />
-    <ClInclude Include="inventorty.h" />
+    <ClInclude Include="item.h" />
     <ClInclude Include="user.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
