--- Data Structures Game/Inventory.cpp
@@ -1,30 +1,0 @@
-//fill out later
-
-
-
-//MISC CODE FOR LATER
-
-//class inventory
-
-
-
-
-
-/*
-void defaultInventory(vector<inventory>& items) {
-	// A vector to store the items
-	// Add medieval-themed items
-	int i = 1;
-	items.emplace_back("Basic Sword", "Made with steel. Its average sized", 1, true, "Weapon", i); i++;
-	items.emplace_back("Basic Shield", "This item wont last that long fr.", 1, true, "Armor", i); i++;
-	//items.emplace_back("Basic Healing Potion", "Medival gatorade.", 5, false, "Consumable", i); i++;
-	items.emplace_back("Gold Coins", "Mideival dolla bill.", 100, false, "Currency", i); i++;
-	//items.emplace_back("Torch", "A wooden torch to light dark areas.", 3, false, "Utility", i); i++;
-	items.emplace_back("Basic Bow", "A simple wooden bow for ranged attacks.", 1, true, "Weapon", i); i++;
-	items.emplace_back("Basic Arrows", "A bundle of arrows for use with a bow.", 20, false, "Ammunition", i); i++;
-	//items.emplace_back("Spellbook", "A tome containing basic magical spells.", 1, true, "Magic", i); i++;
-	//items.emplace_back("Herbs", "A collection of medicinal herbs for crafting potions.", 10, false, "Crafting Material", i); i++;
-	//items.emplace_back("Iron Key", "A heavy iron key that unlocks a mysterious door.", 1, true, "Key", i); i++;
-
-}
-*/
